chore: fix clippy -D warnings across remaining workspace crates (#409)
Mechanical lint fixes (must_use, collapsible_if, while_let_loop,
field_reassign_with_default, too_many_arguments, vec_init_then_push,
doc_list_item, abs_diff, assertions_on_constants, result_large_err,
large_enum_variant, type_complexity, should_implement_trait, manual_map,
nonminimal_bool, if_let_match) across phantom, phantom-adapter,
phantom-agents, phantom-app, phantom-brain, phantom-bundle-store,
phantom-embeddings, phantom-history, phantom-mcp, phantom-net,
phantom-nlp, phantom-plugins, phantom-renderer, phantom-scene,
phantom-session, phantom-stt, phantom-supervisor, phantom-terminal,
and phantom-ui.

`cargo clippy --workspace --all-targets -- -D warnings` exits 0.
`cargo build --workspace` and `cargo test --workspace --no-run` both pass.

+++ crates/phantom-app/src/agent_pane/events.rs
@@ -122,7 +122,7 @@ impl AgentPane {
             .unwrap_or("Sense");
 
         let payload = build_blocked_payload(
-            self.agent.id() as u64,
+            self.agent.id(),
             role_label,
             &reason,
             now_unix_ms(),
